ci: per-surface lint checks (shellcheck / ruff / actionlint)
New one-workflow-per-surface checks, each firing once on main when its
own files change (manual dispatch also available):
- shell-checks.yml: shellcheck on scripts/*.sh
- python-checks.yml: ruff on terraform/lambda/** + scripts/*.py
- yml-checks.yml: actionlint on .github/workflows/**

Ruff fixes so python-checks is green (default ruleset):
- PLW1510: explicit check=False on subprocess.run calls
- UP031/DTZ006/F541: f-strings, tz-aware fromtimestamp, drop stray f
- BLE001: narrow metrics_common catches to (ValueError, TypeError)
- writer: log exc internals, return generic 500 (no leak to beacon)
- I001/RUF046: sorted imports, drop redundant int() cast
- EXE001: chmod +x the four shebang'd python scripts

Also:
- bootstrap-aws.sh: shellcheck SC1007 disable
- terraform.yml: dispatch env option "" -> auto (actionlint empty-string)
- README: document the per-surface check workflows

+++ terraform/lambda/metrics_reader.py
@@ -12,7 +12,6 @@ from datetime import datetime, timezone
 from decimal import Decimal
 
 import boto3
-
 from metrics_common import origin_allowed, request_origin, response
 
 TABLE_NAME = os.environ["TABLE_NAME"]
@@ -102,7 +101,7 @@ def _summary():
                 v = item.get(k)
                 if isinstance(v, (int, float, Decimal)):
                     timing_acc[k] = timing_acc.get(k, 0) + v
-    timing = {k: int(round(v / timing_n)) for k, v in timing_acc.items()} if timing_n else {}
+    timing = {k: round(v / timing_n) for k, v in timing_acc.items()} if timing_n else {}
 
     # Visitors + sessions: group events by anonymous visitor id; a session
     # breaks when the gap between events exceeds 30 minutes.
